REFACTOR : Modification de la structure des dossiers
