+++ src/commands/launch/prompts/search/type.ts
